* Rename web package to view.
