+++ admin/src/main/java/cz/quantumleap/admin/personrole/PersonRoleController.java
@@ -7,8 +7,8 @@ import cz.quantumleap.admin.notification.NotificationService;
 import cz.quantumleap.admin.person.PersonService;
 import cz.quantumleap.core.personrole.transport.PersonRole;
 import cz.quantumleap.core.security.WebSecurityExpressionEvaluator;
-import cz.quantumleap.core.web.DefaultDetailController;
-import cz.quantumleap.core.web.DetailController;
+import cz.quantumleap.core.view.DefaultDetailController;
+import cz.quantumleap.core.view.DetailController;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
