+++ core/src/main/java/cz/quantumleap/core/security/config/MappingInfoRequestMatcher.java
@@ -1,6 +1,6 @@
 package cz.quantumleap.core.security.config;
 
-import cz.quantumleap.core.web.WebUtils;
+import cz.quantumleap.core.view.WebUtils;
 import org.springframework.core.annotation.AnnotationUtils;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.web.util.matcher.RequestMatcher;
