+++ core/src/main/java/cz/quantumleap/core/view/DefaultListController.java
@@ -1,4 +1,4 @@
-package cz.quantumleap.core.web;
+package cz.quantumleap.core.view;
 
 import cz.quantumleap.core.business.ListService;
 import cz.quantumleap.core.common.Utils;
