+++ core/src/main/java/cz/quantumleap/core/view/DefaultLookupController.java
@@ -1,4 +1,4 @@
-package cz.quantumleap.core.web;
+package cz.quantumleap.core.view;
 
 import cz.quantumleap.core.business.LookupService;
 import cz.quantumleap.core.data.entity.EntityIdentifier;
