+++ core/src/main/java/cz/quantumleap/core/view/JsonMapperConfiguration.java
@@ -1,8 +1,8 @@
-package cz.quantumleap.core.web;
+package cz.quantumleap.core.view;
 
 import cz.quantumleap.core.data.entity.EntityIdentifier;
-import cz.quantumleap.core.web.serializer.EntityIdentifierDeserializer;
-import cz.quantumleap.core.web.serializer.EntityIdentifierSerializer;
+import cz.quantumleap.core.view.serializer.EntityIdentifierDeserializer;
+import cz.quantumleap.core.view.serializer.EntityIdentifierSerializer;
 import org.springframework.boot.autoconfigure.jackson.Jackson2ObjectMapperBuilderCustomizer;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.http.converter.json.Jackson2ObjectMapperBuilder;
