+++ core/src/main/java/cz/quantumleap/core/view/ListController.java
@@ -1,4 +1,4 @@
-package cz.quantumleap.core.web;
+package cz.quantumleap.core.view;
 
 import cz.quantumleap.core.data.transport.SliceRequest;
 import org.springframework.ui.Model;
