+++ core/src/main/java/cz/quantumleap/core/view/serializer/EntityIdentifierSerializer.java
@@ -1,4 +1,4 @@
-package cz.quantumleap.core.web.serializer;
+package cz.quantumleap.core.view.serializer;
 
 import com.fasterxml.jackson.core.JsonGenerator;
 import com.fasterxml.jackson.databind.JsonSerializer;
