+++ core/src/main/java/cz/quantumleap/core/view/template/QuantumLeapDialect.java
@@ -1,4 +1,4 @@
-package cz.quantumleap.core.web.template;
+package cz.quantumleap.core.view.template;
 
 import cz.quantumleap.core.filestorage.FileStorageManager;
 import org.thymeleaf.dialect.AbstractProcessorDialect;
